// src/api/openapi/model/schemasOfferV2OfferElementFieldSchemaOfferElementCreatePatch.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SchemasOfferV2OfferElementFieldSchemaOfferElementCreatePatch { 
    defaultValue: string;
    fieldId: number;
    inherits: boolean;
    libraryId?: number | null;
    mandatory: boolean;
}